[Refactor][OPS] Drop dtype from reduction op constructors; derive at forward
Remove the `dtype` ctor kwarg from all reduction ops (3 bases + 11 leaves).
dtype is now read from the input tensor at forward(); the kernel is built
lazily and cached by (cache_key, dtype), and the manifest-derived
_validate_dtypes replaces the inline `x.dtype != committed-dtype` check.
self.dtype is bound from the input before the kernel build, so eval_roofline
stays valid post-forward.

Updates the design playbook (ops-design.md / ops-design-reference.md) to the
new contract: dtype is never a ctor kwarg, dispatch_kernel() stays in __init__,
JIT kernel construction always defers to first forward. Adapts reduction tests
and benchmarks to construct ops without dtype= and key kernel-cache lookups by
(M, N, dtype). Removes two tests that asserted the now-removed
ctor-committed-dtype rejection.

+++ benchmarks/benchmark_base.py
@@ -453,7 +453,7 @@ class ManifestBenchmark(BenchmarkBase[ShapeDtypeWorkload]):
 
     Usage::
 
-        op = SumFwdOp(dtype=dtype, dim=0)
+        op = SumFwdOp(dim=0)
         bm = ManifestBenchmark("SumFwdOp", op, workload)
         result = bm.profile(op, *inputs)
     """
